backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
prisma==0.13.1
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.9
cloudinary==1.39.0
deepface==0.0.86
opencv-python==4.9.0.80
numpy<2
python-dotenv==1.0.1
pydantic-settings==2.2.1
tf-keras
